Alrighty, to my understanding this code is working for the TargetPathingGameObject.cs. This will have the game object move to the target position that it's set to.
If you want this to stop at certain boundings you can configure the bounds and the object will no longer path past it, only to that bound.

Warning: this offsets based on scale so that the game object will set itself to the bounds location with half its scale used to offset accordingly.