I have removed the recently added Player/PlayerClickListener.cs. This was a poor decision, it doesn't accurately handle movement based on object clicks, and for sure doesn't offer the flexibility of responding to a per object interaction.
Created enums to help us start classifying the house items we need to work with and the house carry items to work with. They both extend from HouseBaseObject. This was done so if we need to add functionality between the both of them, we have that available.

What this new click detection does is couple the OnMouseDown method to the house item classes. The house items get a reference to the player. On click they tell the player to move to the position clicked, but to also set the target object of the player to the item clicked.

Not sure what to do next...